[Add] #42 - Task, KeyResult 순서 변경 API

+++ src/main/java/org/moonshot/server/domain/keyresult/dto/request/KeyResultCreateRequestDto.java
@@ -21,7 +21,7 @@ public record KeyResultCreateRequestDto(
         LocalDateTime expireAt,
         @NotNull(message = "KR의 순서를 입력해주세요.")
         @Range(min = 0, max = 2, message = "KeyResult의 순서는 0부터 2까지로 설정할 수 있습니다.")
-        Short idx,
+        Integer idx,
         @NotNull(message = "KR 목표 수치를 입력해주세요.")
         @ValidTargetNumber
         @ValidLimitValue
