+++ src/main/java/org/moonshot/server/domain/task/dto/request/TaskSingleCreateRequestDto.java
@@ -11,6 +11,6 @@ public record TaskSingleCreateRequestDto(
         String title,
         @NotNull(message = "KR 순서를 입력해주세요")
         @Range(min = 0, max = 2, message = "KeyResult의 순서는 0부터 2까지로 설정할 수 있습니다.")
-        Short idx
+        Integer idx
 ) {
 }
